+++ apps/backend/src/modules/users/user.repository.ts
@@ -110,8 +110,43 @@ class UserRepository implements Repository {
 			: null;
 	}
 
-	public update(): ReturnType<Repository["update"]> {
-		return Promise.resolve(null);
+	public async update(
+		id: number,
+		payload: Partial<UserModel>,
+	): Promise<UserEntity> {
+		const user = await this.userModel
+			.query()
+			.withGraphFetched(RelationName.USER_DETAILS)
+			.patchAndFetchById(id, payload);
+
+		return UserEntity.initialize({
+			createdAt: user.createdAt,
+			email: user.email,
+			id: user.id,
+			name: user.userDetails.name,
+			passwordHash: user.passwordHash,
+			passwordSalt: user.passwordSalt,
+			updatedAt: user.updatedAt,
+		});
+	}
+	public async updatePassword(
+		id: number,
+		passwordPayload: { passwordHash: string; passwordSalt: string },
+	): Promise<UserEntity> {
+		const user = await this.userModel
+			.query()
+			.withGraphFetched(RelationName.USER_DETAILS)
+			.patchAndFetchById(id, passwordPayload);
+
+		return UserEntity.initialize({
+			createdAt: user.createdAt,
+			email: user.email,
+			id: user.id,
+			name: user.userDetails.name,
+			passwordHash: user.passwordHash,
+			passwordSalt: user.passwordSalt,
+			updatedAt: user.updatedAt,
+		});
 	}
 }