+++ apps/frontend/src/libs/enums/notification-message.enum.ts
@@ -1,0 +1,5 @@
+const NotificationMessage = {
+	LINK_SENT: "Reset email sent",
+} as const;
+
+export { NotificationMessage };