+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -7,6 +7,8 @@ import { configureStore } from "@reduxjs/toolkit";
 
 import { AppEnvironment } from "~/libs/enums/enums.js";
 import { type Config } from "~/libs/modules/config/config.js";
+import { notification } from "~/libs/modules/notification/notification.js";
+import { storage } from "~/libs/modules/storage/storage.js";
 import { authApi, reducer as authReducer } from "~/modules/auth/auth.js";
 import { quizApi, reducer as quizReducer } from "~/modules/quiz/quiz.js";
 
@@ -19,7 +21,9 @@ type RootReducer = {
 
 type ExtraArguments = {
 	authApi: typeof authApi;
+	notification: typeof notification;
 	quizApi: typeof quizApi;
+	storage: typeof storage;
 };
 
 class Store {
@@ -51,7 +55,9 @@ class Store {
 	public get extraArguments(): ExtraArguments {
 		return {
 			authApi,
+			notification,
 			quizApi,
+			storage,
 		};
 	}
 }