+++ apps/frontend/src/libs/types/types.ts
@@ -1,6 +1,7 @@
 export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
 export { type IconName } from "./icon-name.type.js";
 export { type OPTIONS } from "./input-options.type.js";
+export { type QueryParameters } from "./query-parameters.type.js";
 export { type SidebarItem } from "./sidebar-item.type.js";
 export {
 	type Control,