+++ packages/shared/src/modules/users/libs/types/reset-password-dto.type.ts
@@ -1,0 +1,6 @@
+type ResetPasswordDto = {
+	jwtToken: string;
+	newPassword: string;
+};
+
+export { type ResetPasswordDto };